Image Commit (Terrain Refactor)
Did some major renaming and editing on the terrain folder fixing it so
each terrain is capable of animations as well as defining a naming
convention for dealing with the different terrain types. There are small
comment edits to the LibGDX section of the code as well as the Image
Readme.