Graphic Refactor (Part 1.6)
Not quite ready for Part 2 yet. This commit does some naming
reconventions to the image files. The first part of organization for
sorting the images is complete in the connection code. Finished up the
comments for ObjectStorage.java. Still a lot of debugging work left to
do on the classes, but the Java Client is still stable.