Graphic Refactor (Part 2)
This finishes off the ObjectStorage.java class and further restricts and
changes the naming conventions for the images. Setting the image portion
of the code to "Under Construction", but the graphic data is now
organized into one huge list able to be sorted by the ObjectStorage.java
class. Plenty of changes and debugging to various parts of the code
after a lot of movement refactoring.