Bridge Framework (Json Map)
Well, almost anyway. There seems to be a little bit of a JSON building
issue with the DetachedMapUtil.java class. You can see the problem if
you uncomment the TestMap and run JMain.java and go to Versus (where
this is being tested.)

Other than that, very small changes to the game in general as I prepare
to complete the bridge between the two frameworks. At the moment,
research on how to make a JSON writer and parser until the problem is
resolved.