JSON Parser (Roman Numeral)
JSON to XML conversion finally done without errors. Numerals are not
allowed in XML Tags, so I made a converter that translates all numbers
to Roman Numeral format to handle any number found in a tag. Check out
filler.xml and fillernew.xml to see the difference between the XML
sheets. Now all documents in JSON can be converted seamlessly to XML.

Added MaraSargon to the list of credits for his hard work on the forums.
Thanks!