JSON_Parser (Final Commit)
Finishes off the JSON_Parser allowing for complete parsing of XML and
JSON documents in one class. Edited DetachedMapUtil to create
well-formed documentation for the map. Did some comment editing and
other small code fixes.

Disovered a problem in the Engine Bridge that has to do with how the
system reads maps. Well formed JSON maps are not accepted into the logic
system.