Map Commit (Terrain Display)
Time to start off 2013 with a bang!!!

Terrain display is now available for the Versus Screen in the Java
Client. Images now completely control the index of graphics (Yay!).
Added a missing terrain to the image list. The XML Reader has been
refactored and streamlined to deal with both JSON and XML file types
without errors. Lots of refactoring to all the other classes fixing
numerous bugs and making other classes a little bit more streamlined.