Terrain Display (Testing)
Fixed a few bugs that were present in the first system. Each terrain now
displays properly according to the type given. Roman Numerals were
expanded to deal with zero. Converted the map currently used with the
webClient into JSON so I can perform my tests.

Next on the list is to start on the connection code for terrain.