Sound Commit (AW2 Sounds) - Part 1
A special thank you for JMedeisis for the AW sounds and music. Since
this commit was impossible to contribute in one part due to the vast
size; it is being split up into smaller "byte" size chunks. I changed
the name of __notes.txt to a Readme and renamed the folder, but
everything else about these commits is one-to-one. The first part
uploads the additional folder.

Disclaimer: The sounds and music in the Readme index may not be accurate