Image Commit (Mini-Cannon Recolor and Readme)
This is the first extension of the commit that redoes the readme
explaining the new format of the images. The image renames have not been
done yet, but the format is more or less finalized. I also made the
Mini-Cannons a bit darker for the re-colors. It can be reverted if the
colors do not suit well.