Image Commit (Asset Refactoring)
The mini-cannon re-colors were updated for the inventions section which
will include all future invention images for use within the system. The
mobile folder has been added which will contain drastically shrunk
versions of any assets which need it. Currently, it is just holding
backgrounds, but it can be expanded to hold many different types of
assets.