Asset Commit (Laser Animations) - Part 1
The sheer amount of animations for the laser has forced me to commit
this in parts. Within this part you will find the animations for the
laser charge as well as the weapon fire animation for the laser head.
The "stream" animation is coming as well as the animations for the
mini-cannon (found it!!). Small changes to the read-me.