Asset Commit (Laser Animations) - Final
This commit finished the animations for the laser. The laser animations
are actually a lot bigger than I thought, so pixels will need to be
shifted in order to get the animation to display like in the game.
Otherwise, it'll just be a little off the original animations. Did a bit
of editing to the readme.