Asset Commit (Image Convention) - Part 1
This commit starts the chain of commits that will update all the current
images to the new image convention, and add in the full range of
animation and all the tiles for shoals, sea, reef, and river tiles.

This specific commit moves all the images to a reserve folder. It then
changes all the non-terrain images to follow the new image convention.
Did a bit of changes to the readme and added extra readmes.