Asset Commit (CWT Images)
I have a massive amount of renaming to do for the CWT Images (will
probably automate it). However, the snow example is also here. It is
being tweaked, but the general algorithm is there. The same algorithm
will be used for the rain example, but that'll be in a later commit :).