Map Converter (Breadcrumbs)
Did considerable amounts of work on the map converter and didn't want to
lose the progress, so this commit is just to keep tabs on what has been
done thus far. Did some changes to the Readme and updated the robustness
of the converter class. (Not all folders are added so converter is just
a shell... really).