Music Commit (Tribute to x0_000)
This is all the MIDI's painstakingly created by x0_000 in the past for
the Original Custom Wars. Many were attached to Custom CO's. This commit
was made to preserve the work done. (Also adds a list of authors for the
normal tracks as well as a personal track made for the project, while
deleting redundant copies of songs)