Image Organization - Unit Groups
This is a very simple way of allowing abbreviated unit groups into the
system of Custom Wars Tactics. The groups outlined here serve as a
guideline as I hope when units are put into the system, the groups will
be a lot easier to define. This is a large list holding all the starting
groups.