Image Commit - Unit Dossiers
Due to the inevitable close of Google Code, evasive action had to be
taken to save content. Therefore a wiki was created under
(http://cwt.wikia.com/wiki/Custom_Wars_Tactics_Wiki) to ensure safety of
the documentation. Dossier images used in the wiki were also transferred
to the main site. Colors are not accurate in this version (denoted by
the white background). Some changes were made to the Readme files as
well.