Music Commit - Retsej, the Closing Sonnet
This commit completes the entire collection of the original Custom Wars
tracks. Thanks deeply to Retsej for this contribution as it brings my
frantic search to preserve the Original Custom Wars to a satisfying end.
Now I can actually focus on more current matters with ease. New MIDI and
plenty of MP3 files added to the mix.