Server Commit - Perl User
This is a very rough draft, stripped, and basic overview of the server
layout. A lot of it has commented information, but this commit is just
so I have a working layout of a what a working server skeleton would
look like for CWT.

The main attributes are mostly the commands, which are all documented
very nicely in comments at the top. This version does not work very hard
on the return values, nor does it show layouts for any of the inner files.