Image Commit - Secv90 Snow Properties #2
This finalizes all the major tiles for the snow properties in the normal
games. Since an overlay can be used to simulate rain, it makes all tiles
complete for each weather state. Graphics wise, this game is basically
ready from a map perspective from this commit. Updated all files to
reflect changes.