Pom Pom - The Undoing
Yeah, with all the issues of Windows fixed, the repository is finally
pushed into a stable standing. Now to get some actual work done. Did
some files changes to about everything and fixed up the pom to be more
neutral. Will write up documentation on how to get this running... a bit
later...