Added the ConstructedObject interface. Every property in a constructed class which implements that interface will be instantiated automatically by the factory.
This was also used to extract the logger and the entity manager stuff
from the ConstructedClass and ISystem interface. This decreases the size
of the result classes in the RAM because classes which needs a logger
can bind it by defining a logger property (which will be instantiated
automatically). Before every constructed class got it's own logger
functions, even if they never used them.