JSlix Test Commit - Compression Partisan
This commit adds a bit more to the ImageLibrary class. Even with the
byte compression, the system still has ways to go to hit under a 10MB
limit for images. Preparing to scrape the metal with one of the craziest
plans yet for shrinking the images even further. Also added secv90's
little Partisan unit to the list of units.