JSlix Commit - Learning Phase I
This is pretty much a test to see what I can and can't do with the ST-JS
system. So far, added the JSlix package just trying to mirror the
functionality of the jstest/ImageLibrary.js by getting any image to
display. (I want to see if there is another way without copying one to
one ImgUtil.class already provided). Eventually, it'll happen...