Jslix Commit - Yellow Comet CO Icons
This commit adds the Secv90 Yellow Comet CO Icons to the list of CO
Icons. In other news, ImageLibrary gets an image to appear on the
TempState, but it appears distorted. (You can see it by loading into
TempState if interested. Currently, I'm looking into the causes of why
the image will appear three times)