JSTest Commit - The Big Enchilada
This commit adds a small JavaScript script that takes in text and can
draw an image on command from it. Planning to use this as the framework
for ImgLibrary. More images added in so it was easier to test. Also,
adds in text from Secv90 list of CO units. Next step, getting it to work
with arrays.