Image and CO Icons Update - [AW1 - AW2 - AWDS]
Thanks to the seemingly endless sprite work of secv90, the Commanding
Officer Icons are completed for the first three major games. It is a
tremendous step in working toward perfection in Custom Wars Tactics.
Posted the original secv90 CO Icon artwork, before compression and
resize edits, in a separate folder. In another smaller note, started the
framework for the Image Storage that should deprecate the other one in
time.