Jslix Commit - Balance
This commit fixes a number of issues of the JSTest by adding a queue
system to the images. Altered the FPS a little to count for interval
time in between sleep times. Need to play around with Canvas more to get
more speed out of it. Currently chugging at around 20 FPS with a
background image and one animation, but the good news is that it does
work on all browsers.