JSlix Commit - Eureka
As of this commit, 60FPS has been obtained using the million man march
in all three browsers (Internet Explorer, Firefox, and Chrome). What is
even better, is that external memory usage of the image loader class
will be dramatically reduced. There is a tiny bit of unstable frame rate
in Firefox, but a bit of tinkering should fix that. The new age of CWT
has begun.