JSlix Update - Animation Improvement
Did a small step of getting the image to display and animate via the
JSlix system. It has an arc, but is a little bit faster than the last
iteration. There is still a bit of improvement that can be done to
further speed things up. Also added the statistics for the Custom CO's
by The Thunder.