JSlix Test Commit - Expansion
Got a chance to expand on the ImageLibrary a little bit by adding
functionality for manipulating pixels. The suite of manipulation has
increased to include flipping, rotating, and shifting.

Next up will most likely be finishing the interface so this can be
tested, and adding blending. Promptly followed by memory reduction so it
can be inserted into src_new.