JSlix Test - Basic Functionality
A bit tired., but at least got all the basic functionality in there. All
I have to do is make sure that this is done during the pixel editing
stage and it should be better than the one I made in Java.

The last few steps is optimizing, and making it smart enough to
recognize duplicate items. This should be more than enough to carry
0.36, until a much more powerful screen system is needed. (I'll leave
that for break in case of emergency, and will hope it never gets to that
point.)

The HTML versions have also been edited and expanded to include flipping
the pixels (which makes images appear upside down or left to right upon
clicking a check box. (There is more functionality than that, but I'm
sticking to what looks good with the animations.)