AWDoR/AWDS Commit - Thanksgiving Feast of Sounds
A special Thanksgiving treat of sounds given by Lord Spyro to help round
out the collection of sounds needed for AWDoR and AWDS. Pretty certain
that this completes the collection of sounds needed for the game. As we
already had an extensive AW1 and AW2 collection already. Thanks again
Lord Spyro, for your contributions.