JSlix Test - Refining the Ore
Added some improvements to JSlix. Added functionality for images to be
loaded quicker, rotation, error handling, and mirroring an image. There
is a lot more functionality planned for this, but at the very least it
can be handled in bytesize chunks. Made a pure version in case the new
version gets so slow that it needs to be salvaged.