JSlix Graphics Commit - Prospect Complete
A huge step forward for this engine (jslixrefined.js), it is now
prospect complete. It runs on essentially all current popular browsers
(IE, Chrome, Firefox, Safari, and Opera?) and seamlessly through the
artitectures (PC, Android, and iPhone). Have to remind myself to include
a list of features in the next commit, but the recoloring is the major
update here. It is also a good base stable version that performs pretty
well.