JSlix Refined Alpha - The Final Iteration
Finished a lot of the work associated with this class. It still needs to
be packaged, but I'll cross that river when I get to it. Did a lot of
bug fixing in general, including fixing a coloring mistake within the
building colors (eww). Preliminary testing showed pretty much everything
works, but leaving it in Alpha until a more robust testing suite can be
implemented.