JSlix Commit - The Sound of... Mobile not Working
First expansion into music was a grand success on PC, but just didn't
strike the right chord on mobile. The tunes may be boundless on PC, but
more tuning is required for any such occurence on mobile. MusicTest.html
and MusicLibrary.html hold the music to my ears (on PC only).