JSTest Commit - Portrait of a Landscape
This commit is a placeholder for the tinkering taking place for browser
windows. The progress here is the framework for how the canvas will
perform, but the idea is far from complete. This commit mostly serves as
a bookmark to not lose place on the screen rotation work. Next commit
will hopefully have bigger strides in progress.