JSlix Image Library - The Beginning
Strated converting the old JSlix (jslixrefined.js) into the JavaScript
ES6 (jslix.js). At the moment, looking into why the DOM can't be edited
when this class is built like a module. The one line of code to look at
is 177, which is the most important piece of code of the entire process
because the image can't display without it.