JSlix TextLibrary - ImageLib Fixes
Added the last couple of changes that should let the ImageLibrary do 
everything that I'll need for Text Library to work. The major changes is 
something that allows fonts, the ability to add cut images before and 
after they are manipulated, alongside other changes and bug fixes.

Next step will be finishing the Text Portion of the Image Library, which 
will add a lot of functionality for dealing with paragraphs, lines, and 
words by changing them into easy to manage images.