Jslix TextLibrary - Shortstop
Fininshed the main part which converts typed text into pictures. There 
are a few quality of life features which I placed on the top of the 
class, but with this complete I can leave those for a future date. Next 
on the list of work is getting work done on the actual framework, but 
the work for that is probably going to be on the cwt-server repository 
until we can go back into the planning stage.