Major CWT Commit - The Mind Meld
This commit cements the merging of the desktop client (Java) and the 
online client (JavaScript) into the master branch. Merging the two has 
put the game in a pre-alpha status as the graphical engines will be 
under construction having their functionality matched between both 
architectures.

Javacript: A mix between Tapsi and JSRulz code for higher efficiency and 
flexibility
Java: Completely got rid of Slack2D for a tiny loss of speed, but a more 
stable system

- Credits: Wrote a JSON version and a TEXT version
- MAIN README: Expanded the information
- PIC README: Minor edits
- PICTURES: Added lots and lots of images