CWT Organizational - Text Image Library
This commit begins the work of fully restructuring the code to be a 
machine for development, rather than a hinderance to it. The entirety of 
CWT will be coded at the same time, starting with the the first stop: 
Making a fully functional Graphics library that'll serve the purposes of 
the restructured code. 

Instead of cluttering the Issue List with normal progress, README's will 
be spawned to handle the TO-DO's of certain functions of the project as 
well as organization of notes for that specific section. Hopefully, 
that'll give the motivation needed to really nail down work in the 
project overall.

Let the games begin!