CWT TextonText - Quick Stack
Made a bit of progess on this end, nothing substantial. At least got the 
JavaScript version working on the fly (which is good news) but lost the 
ability to pre-load data (which kind of sucks). Java has to be redone so 
it can do similar work, as it sucks at both on the fly and just doing 
things in general.

- Kind of fixed Jslix Text [JS]
- Updated Testbed [J/JS]
- Updated README [J/JS]